Fixed incorrect query generation in jsonapi_filter
This fixes incorrect queries generated when using multiple filter criteria.
`union_all` combined the results of the subqueries with one filter each applied,
effectively creating an "or" combination of the criteria.
Creating sqlalchemy expressions/critera and feeding them to `filter` creates a simpler query
with an "and" combination of WHERE clauses.